Created Alarm
Showcasing my project that is alram that is created using programming language(python) in this i used basic python fundamentals and also some modules also like tkinter,date, time and winsound. Those modules are imported in the project alarm.
It works when we run the code the desing alarm is displayed in that we will set the time after compled that time the win sound will be ring.